this should fix it (hopefully lol)

+++ packages/backend/dart/test/07_finished_collaborative_documents_test.dart
@@ -20,10 +20,7 @@ void main() {
   });
 
   tearDown(() async {
-    await supabaseAdmin
-        .from('finished_collaborative_p2p_purpose_documents')
-        .delete()
-        .eq(
+    await supabaseAdmin.from('finished_collaborative_documents').delete().eq(
           'collaborator_one_uid',
           firstUserUID,
         );
